Improve via height electrial length calculations
For vias which terminate electrically in an internal layer, only
count half of that internal layer height. This means a subsequent
via on the same route will not result on the internal layer being
double-counted.

Fixes https://gitlab.com/kicad/code/kicad/-/work_items/24118

(cherry picked from commit 848f6dcd2e6cacaad6836ee7ddbd814cec50fcfc)

+++ pcbnew/length_delay_calculation/length_delay_calculation.cpp
@@ -330,7 +330,7 @@ LENGTH_DELAY_STATS LENGTH_DELAY_CALCULATION::CalculateLengthDetails( std::vector
         {
             const auto [layerStart, layerEnd] = item.GetLayers();
             int64_t viaHeight = StackupHeight( layerStart, layerEnd );
-            details.ViaLength += viaHeight;
+            details.ViaLength += static_cast<int>( viaHeight );
             details.NumVias += 1;
             processedVias++;
 
@@ -341,7 +341,7 @@ LENGTH_DELAY_STATS LENGTH_DELAY_CALCULATION::CalculateLengthDetails( std::vector
         else if( item.Type() == LENGTH_DELAY_CALCULATION_ITEM::TYPE::PAD )
         {
             int64_t padToDie = item.GetPad()->GetPadToDieLength();
-            details.PadToDieLength += padToDie;
+            details.PadToDieLength += static_cast<int>( padToDie );
             details.NumPads += 1;
             processedPads++;
